Renamed all models to have proper naming (from model-name.ts to model-name.model.ts), updated includes of services
